DAOS-6522 Security: authentication handshake failed x509
Go 1.15 changed the behavior of CommonName validation for certificates without
a SAN. In order to allow for this change the certificate generating script will
now add a Subject Alternative name (SAN) to the certificates. It also removes
code that forced the client to assume the Server's name was always server.

Signed-off-by: David Quigley <david.quigley@intel.com>