DAOS-5448 Agent: Unexpected default AttachInfo response (#3284)
A race condition was found in the daos_agent when
enough clients simultaneously issued GetAttachInfo.

The agent initialized the attachInfo cache multiple
times due to the race condition.  Some of the cache
init attempts failed to find devices in the fabric
scan due to a second related race condition.  Those
instances that failed to find devices in the fabric
scan populated a cache with only the default entry
and supplied the default entry to the caller.

Two race conditions were fixed, one that controls
the cache initialization, and another that protects
the hwloc library from concurrent access to
hwloc_get_proc_cpubind().

Signed-off-by: Joel Rosenzweig <joel.b.rosenzweig@intel.com>