DAOS-5661 KV: add ops for KV open/close/destroy
Add three new API functions for opening, closing and destroying
a KV object. previosuly KV objects were relying on the basic obj
API for opening and closing, but this add a wrapper around those
for checking for example the oid to make sure its set to flat.

also update the pydaos shim.

Signed-off-by: Mohamad Chaarawi <mohamad.chaarawi@intel.com>