DAOS-5553 dtx: purge cached obj when cleanup DTX handle
For compounded modification, it is possible that some sub-modification
fails as to all the former sub-modifications need to be aborted. Under
such case, the objects (created in former sub-modifications) should be
purged from cache to avoid misguide subsequent operations.

This patch also adds more log messages for DTX related logic.

Signed-off-by: Fan Yong <fan.yong@intel.com>