DAOS-5548 il: Avoid invalid read in interception library.
When closing files in the interception library zero the old
dfs pointer after unmount, so that when subsequent files are opened
the container is remounted.  This avoids reading from freed memory
which was causing unstable behaviour and performance problems.

Correct an error number on error handing in open, and improve log
parsing in some cases.

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>