DAOS-4633 vos: Propagate punches when result is empty tree
In order to avoid emptiness checks on iteration, we can instead
propagate punches that result in an empty tree to the parent
key/object.   Thus we pay a penalty on punch but make it up
on update.

Adds an internal iterate function based on an open tree handle
to do the emptiness check.
Disables this behavior temporarily for empty object as it breaks
some existing rebuild tests in the absense of a fix for DAOS-4698.
Until this is fixed, we don't support emptiness check on vos
objects.
Enable the tests that were failing emptiness checks.
Enable immediate abort in mvcc tests if CMOCKA_TEST_ABORT is set
and one of the cases fails

Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>