DAOS-5048 Client: Improve GetAttachInfo error reporting (#3476)
If the libdaos client receives an invalid GetAttachInfo
payload, it now returns -DER_AGENT_INCOMPAT if the payload
is missing key components, or -DER_INVAL if a component of
that payload is malformed.

-DER_AGENT_INCOMPAT means that there is version skew between
the libdaos.so client library and the DAOS Agent.  The DAOS
Agent may be older than the client library.  In this case,
the remedy is to update the DAOS Agent to sources that are
compatible with the client library.  To ensure compatibility,
use the libdaos.so, DAOS Agent, and DAOS Server from the same
build.

Signed-off-by: Joel Rosenzweig <joel.b.rosenzweig@intel.com>