686 CART: Reuse of OFI port number should return error.
CART-686 test:
When an OFI_PORT is used and another server instance tries to
open the same port, than crt_context_create should return an error.
For sockets, this is not true (but sockets are no longer supported).
TCP does return an error.
We added a unit test to verify that the provider return an error
if necessary.

Signed-off-by: Steven Bollinger <stevenx.bollinger@intel.com>