DAOS-5435: weekly-testing: IorIntercept: test_ior_intercept: IOR command issued warnings
'ior' command uses the -D optional argument to specify the number of seconds
before stopping a read or write phase. If for some reason the data written or read
takes longer than the seconds allowed by this parameter the test logs WARNING
messages and finally fails the test.

Being exposed that, an optional boolean parameter named 'fail_on_warning' was introduced
in the run_ior_with_pool() function to control whether the test should fail or not.
In order to test the correct behavior of this parameter, the test 'ior_fail_on_warning.py'
was added too.

Test-tag: iorfailonwarning
Test-tag-hw-small: iorfailonwarning

Signed-off-by: David Maldonado Moreno <david.maldonado.moreno@intel.com>