DAOS-5748 test: fix parsing of dmg pool create svc replicas
Change the regex used to parse the dmg command standard output
to gather the pool service replica ranks and construct an array
that is used in the pydaos class DaosPool API implementation.
Before this change, only the first pool service replica rank
was parsed. With this change, all replica ranks are parsed from
the comma-separated list output.

Signed-off-by: Kenneth Cain <kenneth.c.cain@intel.com>