DAOS-5754 rsvc: RPC error processing for membership changes
A service membership change (such as removing a replica) may result in the
new service leader briefly replying with the -DER_NOTLEADER error along
with a hint referring to itself as the next leader. This is due to a race
on the server side between raft protocol transition to leader, and bringing
the rdb_raft state to UP. Clients processing RPC replies need to be prepared
for this and ultimately understand the identity of the new service leader,
to avoid always performing a "find the leader" RPC loop across multiple
replicas (leader and standbys).

This change augments the rsvc RPC reply error processing to accept
a hint argument that is sometimes available such as when a new leader
replies NOTLEADER with a hint. This allows the function to distinguish
between an old leader stepping down, and a new leader about step up.
Hint processing is also enhanced to make sure a new leader stepping
up has its aliveness value set/maintained appropriately to avoid
a search for the leader loop.

Signed-off-by: Kenneth Cain <kenneth.c.cain@intel.com>