DAOS-5818 test: Fix NLT test_alloc_fail to use real pool, and resolve errors.
The fault injection test in NLT was using a hard-coded uuid
instead of a valid one so correct this.  This means we do
not get any NOTREPLICA errors from the server which is good,
but does mean we have another couple of leaks from the client,
so fix these.

Add re-try logic around pool_disconnect so even under NLT
it manages to complete correctly, othewise without being
able to disconnect from the pool references are held and
a considerable amount of memory is leaked.

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>