DAOS-5754 rsvc: RPC error processing for membership changes (#3580)
A service membership change (such as removing a replica) may result in
the new service leader briefly replying with the -DER_NOTLEADER error
along with a hint referring to itself as the next leader. This is due
to a race on the server side between raft protocol transition to
leader, and bringing the rdb_raft state to UP. Clients processing RPC
replies need to be prepared for this and ultimately understand the
identity of the new service leader, to avoid always performing a
"find the leader" RPC loop across multiple replicas.

This change augments hint processing to make sure a new leader stepping
up has its aliveness value set/maintained appropriately to avoid a
search for the leader loop. This hint processing can counteract
the "give up this leader" logic that will have just happened in the
preceding rsvc_client_process_error() function in this scenario.

Signed-off-by: Kenneth Cain <kenneth.c.cain@intel.com>