DAOS-5874 tests: Report errors for tests not executed.
Launch.py does not execute any remaining tests in the list when a
communication error is detected for one of the hosts under test prior to
starting the current test.  This was not clealy being reported to
Jenkins and the lack of test execution could be missed.

The launch.py script will now generate a JUnit xml file for each test
which was specified but it does not execute.

Signed-off-by: Phillip Henderson <phillip.henderson@intel.com>