Lock the membership on group update, not the database
This really needs to be reworked. Locking the database was
not enough, because the membership.Join() method only takes
short-term locks on the Database, leaving ample opportunities
for races. Locking the membership on group update should prevent
a concurrent Join from occurring with the group update, which can
result in the group map not being updated with the joining rank.