Updating the get_test_list() method.
If launch.py is called with a combination of test tags and one or more
test files it will now apply those tags to the test files to determine
if they should be run.   This will avoid reporting a test as not run
due to tag filtering by the 'avocado run' command.

Signed-off-by: Phillip Henderson <phillip.henderson@intel.com>