DAOS-5711: seg fault in client library
A Segmentation Fault happens when calling daos_obj_open() without initializing
daos previously.

The "__gurt_list_add()" function inserts a new element into a linked list
passed as a d_list_t head parameter.
The Seg Fault occurs, in this particular scenario, because the list is empty
since it was not initialized before trying to insert into it.

The initialization of this list (scheduler's dsp_running_list) is performed
when calling the function "daos_init()".

In order to prevent this seg fault behavior, a validation was added to the
daos_event_init() function, as well as a new DAOS error "DER_EQ_UNINIT" with
description: "Event queue is not initialized".

Signed-off-by: David Maldonado Moreno <david.maldonado.moreno@intel.com>