DAOS-5731 md: rdb/raft regression test Coverity defect fix
The rdbt client code in rdbt_find_leader() erroneously issued a break
that exits the containing for loop. This was done when a sanity check
on an RPC reply failed. The intent was to goto a test near the bottom
of the loop to print an error and set a return code indicating error.
The test at the bottom of the loop was effectively dead code, detected
in a coverity scan of the daos code.

With this change, the initial tests are converted to gotos and the
previously dead code is now reachable. Also with this change, a typo
in the rdbt client command line usage is fixed (--nranks argument).

Skip-unit-test: true
Skip-func-test: true
Skip-func-hw-test: true

Signed-off-by: Kenneth Cain <kenneth.c.cain@intel.com>