DAOS-623 build: Fail the build if no commit version is specified
For prerequisites, we should not rely on stability of upstream
commits.   This patch fails the build if nothing is specified
as a commit version in utils/build.config for a component
we clone.   This still allows someone to locally specify
something bad like "master" but I think this is potentially
useful for local testing so I leave it as a possibilty and the
responsibility of the gatekeeper to avoid landing.

Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>