DAOS-5571 dtx: dtx guard epoch
Anytime when the server (re)start, the dtx_guard_epoch will be set as
current known highest HLC. In theory it should be the highest one for
the whole system, any other transaction with old epoch (HLC) in spite
of being generated by which server will be regarded as started before
current server (re)start. Current server will refuse such transaction
and require its sponsor to restart it with newer epoch.

On the other hand, during DTX recovery, if the DTX on some server only
contain read operations, then the leader will specify the DTX epoch in
the DTX check RPC ('di_epoch'), and then related non-leader will check
whether it is newer than the dtx_guard_epoch on current server, if yes,
regard the DTX as prepared in spite of whether such DTX has been really
handled on this server or not before current server the latest restart,
because the guard epoch on the server will refuse any modification with
old eopch, that have the same effect as related read operations for the
DTX have been executed on current server.

Be as a temporary solution, set the dtx_guard_epoch as current server
local HLC when (re)start plus the hlc epsilon. If the hlc epsilon is
maintained well, then the dtx_guard_epoch may be a bit higher than the
theoretical system level highest HLC at that time.

Signed-off-by: Fan Yong <fan.yong@intel.com>