DAOS-4928 dfuse: Remove inode record table from dfuse.
Overhaul how inodes are allocated in dfuse, rather than allocating
an integer and searching that it's unique make a inode from the
oid and dfs, and then check that it's unique.

This removes the inode record table entirely, which was a record
of every inode used ever, where all lookups were negative and
uses just the inode entry table which was searched anyway, there's
still a find_insert() call for each inode here.

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>