DAOS-5563 tests: Run find(1) on a dir tree created with posix containers
This test creates and mount a number of POSIX containers and link them
together such that they form a tree with a common root directory.
Then, the test runs find(1) to locate specific files in various parts
of the tree.

If the optional “challenger_path” parameter is provided, the test will
perform the same operations on such path and compare the execution time
against DAOS.

Signed-off-by: Jonathan Martinez Montes <jonathan.martinez.montes@intel.com>

Test-tag: findcmd