DAOS-5571 common: server start epoch
Anytime when the server (re)start, the dss_start_epoch will be set as
current known highest HLC. In theory it should be the highest one for
the whole system, any other transaction with old epoch (HLC) in spite
of being generated by which server will be regarded as started before
current server (re)start. Current server will refuse such transaction
that contains modifcations and require its sponsor to restart it with
newer epoch.

Here is about how to generate the server start HLC (epoch). If server
restart so quickly as to all related things are handled within the HLC
epsilon, then it is possible that current local HLC on the server after
restart may be older than some HLC that was generatee before the restart
because of the clock drift between servers. To prevent that, we control
the server start speed to guarantee that the restart time window will be
longer than the HLC epsilon, then new HLC generated after server restart
will not rollback.

Signed-off-by: Fan Yong <fan.yong@intel.com>