DAOS-5726 vos: Fix a bug with subtree creation
If a parent tree is created in a PMDK transaction such
as a punch where a subtree is not created, a subsequent
transaction that creates a subtree will modify the krec.

If this happens, the krec is not added to the PMDK
transaction.   This can cause a subsequent abort to leave
garbage in the krec for either the incarnation log or
the bmap flag and get the tree into an unusable state.

Adds a test case to try to stress VOS paths.  This was
an attempt to reproduce the issue in the ticket.  In doing
so, I discovered this bug.

Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>