DAOS-4709 dmg/bio: Add dmg storage replace command for hotplug feature
$dmg storage replace --old-dev=UUID --new-dev=UUID
This command will replace the "old-dev" with the "new-dev". This is apart
of the NVMe Hotplug feature, and assumes that the device being replaced
is a hot-removed device being replaced by a newly plugged device. In theory,
the old device should stay as an "UNPLUGGED/OUT" state, and the new device
should go from a "NEW" state to a "NORMAL" state, after the SSD has been
reintegrated.

This command can also be used to reuse a FAULTY device (mainly for test
case purposes). If a device is manually set as FAULTY via
$dmg storage set nvme-faulty, it can now be transitioned back to a normal,
usable device by setting "old-dev = new-dev = DEVICE_UUID".

The command will print the new device UUID and the state on the console,
similar to the set faulty command.

Signed-off-by: Sydney Vanda <sydney.m.vanda@intel.com>