DAOS-5735 - All pool bad_evict tests failed on weekly due to DaosApiError: No existing UUID on pool.destroy
Different updates done in the test in order to remove
the dependency of the DaosPool.create() method, since
when it was removed from pydaos this test started to
fail.
Now the test uses the DmgCommand for pool creation and
pool eviction.

I also removed from the yaml file the different scenarios
of pool create, first because the intention of the test
was to excercise eviction, and second because the pool
create operation with dmg accepts any DAOS system name
passing it with the --sys flag/option.

Test-tag: pr,-hw badevict

Signed-off-by: David Maldonado Moreno <david.maldonado.moreno@intel.com>