DAOS-5657 vos: Add epoch uncertainty checks to VOS (#3879)
Adds epoch uncertainty checks to VOS
- adds write timestamps to the timestamp cache
- Add uncertainty tests to vts_mvcc.c

Two write timestamps are added.  Otherwise, we would
have to reject every operation that has a later write
regardless of whether it's inside the uncertainty
bound.

Note, this can potentially be improved a bit for
query and list cases as the checks at present
are done at a higher level than necessary.  For
example, for object listing, the check is done
on the container rather than each object.  This
should cover the objects but is a bit of a
heavy hammer.   But it's a line in the sand.

A few other changes:
- No need to pass dth to vos_iterate for punch propagation since we
already have relevant keys to update
- Fix a logging issue
- Fix a bug causing timestamp updates on tx failure

Co-authored-by: Li Wei <wei.g.li@intel.com>
Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>