DAOS-6087 il: Handle large pool sizes by backing off.
When a global pool handle is too large then back off to using
pool_connect etc rather than aborting.  This works, but does mean
on large systems additional steps are required to setup
interception, so an attempt at compression would be preferred here.

This re-instates the behaviour that we had until
4da39276479affaf05d79cd6010ac0142cacc016 where the IL would retry
on any error, the improvements there included only retrying on EPERM,
but this change allows it to retry on EOVERFLOW as well.

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>