DAOS-5703 dfuse: Add mknod call and use special files for UNS.
Rather than using directories for UNS entry point use the
mode bits for FIFO.
Change dfs to accept FIFO as a mode, and to treat them as
directories, but not allow readdir or creating files.
Change dfuse to allow creation of FIFOs, but once they're
created then report them as dirs.

Resolve a bug in NLT where cleanup of agent dirs was happening
early, leading to errors on failure paths.

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>