DAOS-4711 dmg/vmd: Add dmg storage identify command for VMD LED management
**ERROR- need to debug timing duration for LED state (currently resetting
before 60 sec have occurred) **

This functionality is apart of the NVMe Hotplug feature and allows for
SSD identification with the use of VMD LED device management.

$dmg storage identify --traddr --uuid command with the specified transport address
and device UUID will now set the status LED state on the VMD device to an "IDENTIFY"
state. This will be visually represented by a quickly blinking amber LED for approx
5 minutes, and then the LED will return to the default OFF state.

**NOTE: This capability is only available on VMD devices (i.e. VMD functionality
needs to be enabled in the system BIOS, and also enabled in DAOS server config)
	disable_vmd: false
	bdev_class: nvme
	bdev_list: ["0000:5d:05.5"]

The transport address for the VMD device can be found by querying the devices
(i.e. running $dmg storage query list-devices).
    UUID:43496b73-7b29-4613-8105-64f08bb4bc38 [TrAddr:5d0505:01:00.0]
      Targets:[0] Rank:0 State:NORMAL

Signed-off-by: Sydney Vanda <sydney.m.vanda@intel.com>