DAOS-623 build: Add a BuildPriority parameter
To make use of the Priority Sorter plugin.

This requires you to go to the Classic View page for your PR and click
Build with Parameters and then fill in a priority for your PR.  That
will cancel the existing run-in-progress and start a new run with the
specified priority.

This is a feature that should be used sparingly and only in the most
emergent cases of needing to prioritize a PR.  My suggestion would be
that it requires gatekeeper approval to use.  But use pollicy is some-
thing that can be decided later and tuned as time goes on.

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>