DAOS-6133 rebuild: store migrated objects into migrated tree
Since duplicating write with the same epoch might cause some
trouble for VOS, so let's remember all migrated objects in a
separated migrated tree to avoid the objects being migrated
multiple times.

We used to have this in the old rebuild code, but got removed
when separating the migration with rebuild.

Signed-off-by: Di Wang <di.wang@intel.com>