DAOS-5989 tools: Include ACL in daos cont get-prop (#3921)
Update the daos cont get-prop command to include the ACL
prop if the user has access to it. For now this means
we make two API calls: one to daos_cont_query excluding
the ACL, and one to daos_cont_get_acl. If the user isn't
allowed to access the ACL, nothing about the ACL prop is
printed.

Signed-off-by: Kris Jacque <kristin.jacque@intel.com>