DAOS-6160 control: Fix detectVMD() output to add address prefix when missing
When multiple devices are listed in PCI_WHITELIST to be unbound by SPDK,
the VMD address that is queried by $lspci | grep "Volume Management Device|201d"
loses it prefix of "0000:" which results in DAOS not using the correct
backing transport address for the SSDs behind the VMD address.

Add check for this prefix, and include it when missing.

Signed-off-by: Sydney Vanda <sydney.m.vanda@intel.com>