Update based on discussion
After discussion with Ken & Johann, I've adjusted the retry
logic a bit. First of all, we shouldn't retry pool creates
with a new UUID, as this can lead to a variety of undesirable
scenarios (e.g. spurious -DER_NOSPACE). Instead, we should
enforce that a single pool create should always be associated
with a single UUID. Also add -DER_AGAIN to the set of retryable
creation errors, and catch a -DER_EXIST in the target
creation code in order to turn it into a -DER_AGAIN.

Signed-off-by: Michael MacDonald <mjmac.macdonald@intel.com>