DAOS-5900 control: Add support for unique pool labels (#3974)
Expands MS DB to include a map of label->UUID in order
to enforce unique labels within a system. Adds a new
PoolResolveID() RPC to allow control clients to resolve
a "human friendly" ID (e.g. a label) into a UUID for use
with other API calls. Updated the dmg command surface
to accept a UUID or a label with the --pool parameter.

Also fixes some issues with the earlier property setting code
which, among other problems, didn't handle clearing a label
correctly.

A subsequent PR will add support for libdaos applications to
use non-UUID identifiers when interacting with a pool.

Signed-off-by: Michael MacDonald <mjmac.macdonald@intel.com>