DAOS-6100 vos: Update negative timestamp cache (#3910)
Rather than requiring a positive entry exist, make the
negative cache global. At each level, the index used
is based on the hash of the object/key and the path or
the parent object/key. These are accessed directly.

Consider the following scenario which illustrates
the issue with the prior implementation:

T1 does conditional insert of dkey1 and T2 does
conditional insert of dkey2 on the same object.
This gets split in a distributed transaction into
a conditional fetch followed by an update on commit.
The issue with the old implementation happens when
the object shard does not yet exist. Both fetches
would update the read timestamp on the negative
object entry.

This patch allows us to check and update
a likely unique negative dkey timestamp instead
which should significantly reduce these types of
conflicts.

It does not eliminate the possibility of false
conflicts but they should be extremely rare.

Also changes rebuild to avoid duplicate akeys

Co-authored-by: Di Wang <di.wang@intel.com>
Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>