DAOS-4711 dmg/vmd: Add dmg storage identify command for VMD LED management (#3950)
This functionality is a part of the NVMe Hotplug feature and allows
for SSD identification with the use of VMD LED device management.

Running `dmg storage identify vmd --uuid=DEV_UUID` with the specified
device UUID will now set the status LED state on the VMD device to an
"IDENTIFY" state. This will be visually represented by a quickly
blinking amber LED for approx 1 minute, and then the LED will return
to the default OFF state. We can make this LED period configurable
from the command line in a future improvement.

**NOTE: This capability is only available on VMD devices, which means
that VMD functionality needs to be enabled in the system BIOS and also
enabled in DAOS server config, an example of which follows:

disable_vmd: false
bdev_class: nvme
bdev_list: ["0000:5d:05.5"]

Output of the command looks something like the following:
Devices
  UUID:4843aa4a-bb88-48be-b47e-89e760d83d11 [TrAddr:]
    Targets:[] Rank:0 State:IDENTIFY

The "State" returned here is the state of the VMD LED that just got set,
not the state of the device. Transport address and targets will remain empty
for this command response.

If a non-VMD device UUID is used, the output will look like the following:
Errors:
  localhost DAOS error (-1010): DER_NOSYS

Signed-off-by: Sydney Vanda <sydney.m.vanda@intel.com>