DAOS-2666 task: the task struct should not use const for uuids
Users using the task API should be able to set the uuid in the struct
for operations that take them as input parameters (pool/container
create, destroy, etc.). This is different than an API call where
the uuid is passed vs being set in a struct which will cause warnings
in the user code.

Signed-off-by: Mohamad Chaarawi <mohamad.chaarawi@intel.com>