DAOS-5479 dmg/vmd: Set FAULT state on VMD LED when device is evicted (#4000)
This functionality is a part of the NVMe Hotplug feature and allows for SSD
identification with the use of VMD LED device management.

When a device is manually evicted via the `dmg storage set nvme-faulty` command,
if the device is a VMD-enabled NVMe SSD, then the FAULT state will be set
on the device LED. This will be visually represented by the LED being turned to a
constant "on" state. This LED state will persist on the device which will
need to be replaced by a new device (i.e. hot-plugged).

Signed-off-by: Sydney Vanda <sydney.m.vanda@intel.com>