DAOS-6149 EC: set iod_size by reply size for EC object
Add orr_size_set to indicate if the iod_size has been
set by fetch shard reply, and only update size if
iod_size is not being set or iod_size is still 0.

After fetch the extents from parity node, let's
check if the extents needs to be rebuilt, i.e. if
the extents located on the rebuilding shard, and
skip the iod if no extents needs to be rebuilt.

Signed-off-by: Di Wang <di.wang@intel.com>