DAOS-6227 vos: set epoch/pm_ver as lowest when coalesce exts
To fix the issue met in EC degraded fetch, set the merged ext as the
lowest epoch of the to-be-merged exts.

A few details to explain the issue, for example for a 2+1 EC obj, cell
size is 1MB. Firstly write with two full-stripe updates – [0-2M] and [2-4M].
Then with vos exts [0-1M]@epoch1, and [1-2M]@epoch2 on all data+parity
shards(shard 0, 1, 2).
Now data shard 1 dead, and user fetch [1-2M], the degraded fetch req
will be sent to parity shard 2, it checks parity/replica space and
replies client need to recover from [0-2M]@epoch1, and client will try
to read back VOS ext [0-1M]@epoch1 from shard 0 and 2. Here vos
aggregation cut in and possibly merge vos exts on shard 0 and 2, in that
case if the new merged ext use epoch2, it will cause the client get the
incorrect data for recovery.

Signed-off-by: Xuezhao Liu <xuezhao.liu@intel.com>