DAOS-5696 Agent: Enable Agent to communicate to the Control Plane to clean up handles
When the agent detects that a process has closed without disconnecting its
handles or has died before being able to do so the agent needs to relay this
information to the control plane. A new gRPC PoolHandlesCleanup is used to
relay to the control plane the pool UUID and associated handles to clean up.

Signed-off-by: David Quigley <david.quigley@intel.com>