DAOS-623 build: Enable avx2 in build
The DESIRED_FLAGS entry for -mavx2 had a space which causes some
random behavior.  On my box and on boro, it causes it to fail the
configure check even if the option is available.  Neil tried
building on his OpenSuse laptop and it passed the configure
check but then failed to compile anything because it passed
the option as " -mavx2" rather than just -mavx2.

Removing the space fixes it and probably enables the extention
on most systems.

Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>