DAOS-6384 control: Clean up gRPC/protobuf namespaces
The distinction between the mgmt/ctl namespaces has gotten
blurry over time as new RPCs have been added to each namespace.

This commit brings the namespaces back into alignment with the
original layout, which is that the mgmt namespace is for RPCs that
interact with the Management Service (running on hosts listed in
the access_points server config parameter) and ctl RPCs may be
invoked on any control plane server (e.g. StorageQuery,
StorageFormat, etc.)

Signed-off-by: Michael MacDonald <mjmac.macdonald@intel.com>