DAOS-XXXX ras: internal io_server interface
Just a sketch to get the discussion started on the internal API
for RAS event in the I/O server. The code uses inline for
convenience, but we should build a real API using drpc under
the hood too communicate the RAS event to the control plane.
It was also requested that RAS events are logged to the
DAOS debug logs so that developers can correlate those events
with other debug messages. This patch just prints RAS event
to stdout, which isn't what we want.

Feedback is welcomed.

Skip-test: true
Skip-build: true

Signed-off-by: Johann Lombardi <johann.lombardi@intel.com>