DAOS-6238 dfuse: Ensure inode object is opened with O_RDRW. (#4166)
* DAOS-6238 dfuse: Ensure inode object is opened with O_RDRW.

This means that the inode always has write permissions, even if the
open file derived from it may not, and it fixes a bug where files
were created or opened in read-only mode.

Mkdir also needs updating however I have another fix for mkdir in review
so I will add that change there.

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>