DAOS-6396 control: CLI utils should exit 0 after --help
The go-flags package builds the help message in an error which
was printed in the exitWithError() helper, but this is potentially
confusing because it would result in the utilities exiting with 1
instead of 0. Just inspect the error returned from parseOpts() and
only pass it to exitWithError() if it's actually an error.

Signed-off-by: Michael MacDonald <mjmac.macdonald@intel.com>