DAOS-5696 control: Add agent -> server communication for stale handle cleanup (#4142)
When the agent detects that a process has closed without disconnecting its
handles or has died before being able to do so the agent needs to relay this
information to the control plane. The PoolEvict command has been extended to
take an optional set of handles to evict without evicting the entire pool. This
change enabled the leaked handle detection and modifies the PoolEvict call to
the control plane. However the PoolEvict call into the io server does not yet
make use of this handle list for performing selective eviction.

Signed-off-by: David Quigley <david.quigley@intel.com>