DAOS-6425: Store UT Valgrind XMLs in current directory (#4220)
When there is a failure on the CI UT with memcheck stage, test_main_node.sh
exists because it has set -e. This prevents this script from moving the
Valgrind XMLs to current directory where pipeline-lib scripts can find them
and archive them for later debug of the memory leaks.

In practice, it is not necessary to have those Valgrind XMLs in test_results
and then move them. We can reduce this step and overcome this issue by
directly keeping those files in the current directory.

Signed-off-by: Marcela Rosales <marcela.a.rosales.jimenez@intel.com>