DAOS-6357 test: Fix NVMe Recovery 3 test failure
Currently NVMe Recovery 3: Verify blobstore state NORMAL->OUT transition
test was failing when ran on more than 1 node. This was due to the fact
that the device being set as FAULTY and then queried for the device state
may not have been present on the server and then unable to be found in
the per-server metadata. The test was updated to use a device only on
rank 0 to avoid this error.

Signed-off-by: Sydney Vanda <sydney.m.vanda@intel.com>