DAOS-6187 mgmt: Return all ranks in GetAttachInfo (#2854)
The MS will give the URIs of all ranks it is aware of at the moment of
the GetAttachInfo request, instead of only those of the MS ranks. It
will additionally indicate which among them are the MS ranks.

[protobuf] Rename GetAttachInfoResp's psrs field to rank_uris.
Exposing the crt PSR concept at this level was a mistake (of this
author).

[protobuf] Add GetAttachInfoResp's ms_ranks field to indicate
which ranks are local to the MS replicas. This is currently
required by dc_mgmt_get_pool_svc_ranks; it might also be useful
for other purposes in the future.

[protobuf] Fix the protobuf naming style of
GetAttachInfo{Req,Resp} fields. Following the official standard
(e.g., "ms_ranks" instead of "msRanks") enables both the Go and
the C protobuf compilers to generate good field names in their
languages. We should fix this before interoperability becomes a
hard requirement.

[Go] Change the logging of GetAttachInfo responses to only record
one, instead of all, rank URIs, as the number of all the rank URIs
may be too large for appearing twice per GetAttachInfo RPC in the
debug logs.

[Go] Check GetAttachInfoReq.Sys in daos_agent up front. This is
required to fix evict_test.py's invalid system name case. The test
case used to pass because of the group name check during crt URI
lookups, which may not happen (with or without this patch). The
check is skipped for the special daos_init-dc_mgmt_net_cfg case.

[C] Specify an empty system name, instead of "daos_server", in the
daos_init-dc_mgmt_net_cfg case. When we have multi-system support
in the future, it would be hard to know if a GetAttachInfoReq with
system name "daos_server" is a valid request from a legacy client
or an invalid request from an up-to-date client, if "daos_server"
is unknown to the daos_agent.

[C] Remove all but the system name field from sys_buf to reduce
global pool handle sizes. All client processes, regardless of
whether they utilize l2g2l, will make GetAttachInfo dRPCs to their
local daos_agent processes.

[C] Fix an issue in dc_mgmt_net_cfg where we call put_attach_info
when the get_attach_info call fails.

[C] Remove a couple of if statements whose conditions can never be
true.

Signed-off-by: Li Wei wei.g.li@intel.com